split admin product to components
